specify triangles array dtype explicitly as int
otherwise it defaults to float and it can't be used to index other arrays

+++ tetgen/pytetgen.py
@@ -155,7 +155,7 @@ class TetGen(object):
     @property
     def mesh(self):
         """Return the surface mesh"""
-        triangles = np.empty((self.f.shape[0], 4))
+        triangles = np.empty((self.f.shape[0], 4), dtype='int')
         triangles[:, -3:] = self.f
         triangles[:, 0] = 3
         return pv.PolyData(self.v, triangles, deep=False)
